delay

+++ LIB/PPPC/inc/MISC_def.h
@@ -4,12 +4,13 @@
 #include "stm32f10x.h"
 #include "stdio.h"
 #include "stdlib.h"
+#include "stm32f10x_it.h"
 
 /*关于延时函数*/
 namespace delay
 {
 	void Init();
-	void delay_ms(u16 nms);
+	void delay_ms(u32 nms);
 	void delay_us(u32 nus);
 }
